saoe-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saoe-core"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "PyNaCl>=1.5.0",
    "jsonschema>=4.23.0",
    "Pillow>=10.4.0",
    "markdown>=3.7",
    "bleach>=6.1.0",
]

[project.scripts]
saoe-publish-template = "saoe_core.publisher:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["N802", "N803"]
